changed signature of Persister::some to accept String,Int,Int

+++ src/main/scala/basic/persistence/Persister.scala
@@ -7,5 +7,5 @@ trait Persister {
 	def get(table:String,id:Serializable):Savable
 	def save(obj:Savable):Serializable
 	def search(table:String,field:String,value:Any):List[Savable]
-	def some(table:String,count:Int,offset:Long):List[Savable]
+	def some(table:String,count:Int,offset:Int):List[Savable]
 }
